refactor: 패키지 변경, 타입 변경, 주석 추가 (#3)
