--- src/main/java/prography/pingpong_game/room/Team.java
@@ -1,5 +1,0 @@
-package prography.pingpong_game.room;
-
-public enum Team {
-    RED, BLUE
-}
