+++ src/main/java/prography/pingpong_game/room/entity/RoomStatus.java
@@ -1,4 +1,4 @@
-package prography.pingpong_game.room;
+package prography.pingpong_game.room.entity;
 
 public enum RoomStatus {
     WAIT, PROGRESS, FINISH
