+++ src/main/java/prography/pingpong_game/room/entity/RoomType.java
@@ -1,4 +1,4 @@
-package prography.pingpong_game.room;
+package prography.pingpong_game.room.entity;
 
 public enum RoomType {
     SINGLE, DOUBLE
