+++ src/main/java/prography/pingpong_game/room/entity/Team.java
@@ -1,0 +1,5 @@
+package prography.pingpong_game.room.entity;
+
+public enum Team {
+    RED, BLUE
+}
